Update dev dependencies for and builds of custom CSS code
In superproject's CSS dev dependency ta-custom_daesa.less, add style rules for hiding an image attachment that is being erroneously introduced by a bug to pages with an HTML snippet and JSON feed in excerpt mode. Next, update the WSU-UE---CSS submodule, which contains universal dev dependencies used to build custom CSS code in every DAESA website development project. Finally, rebuild the files containing custom CSS code that is ready to be deployed on the website associated with this project.